+++ src/dispatching/mod.rs
@@ -1,4 +1,5 @@
-//! Updates dispatching.
+//! Old updates dispatching (**DEPRECATED**: use [`crate::dispatching2`]
+//! instead).
 //!
 //! The key type here is [`Dispatcher`]. It encapsulates [`Bot`] and handlers
 //! for [all the update kinds].
@@ -45,11 +46,12 @@
 //! [`tokio::sync::mpsc::UnboundedReceiver`]: https://docs.rs/tokio/0.2.11/tokio/sync/mpsc/struct.UnboundedReceiver.html
 //! [examples/dialogue_bot]: https://github.com/teloxide/teloxide/tree/master/examples/dialogue_bot
 
+#![allow(deprecated)]
+
 pub mod dialogue;
 pub mod stop_token;
 pub mod update_listeners;
 
-#[cfg(feature = "ctrlc_handler")]
 pub(crate) mod repls;
 
 mod dispatcher;
@@ -63,7 +65,13 @@ pub use dispatcher_handler_rx_ext::DispatcherHandlerRxExt;
 use tokio::sync::mpsc::UnboundedReceiver;
 pub use update_with_cx::{UpdateWithCx, UpdateWithCxRequesterType};
 
+#[cfg(feature = "dispatching2")]
+pub(crate) use dispatcher::{
+    shutdown_check_timeout_for, shutdown_inner, DispatcherState, ShutdownState,
+};
+
 /// A type of a stream, consumed by [`Dispatcher`]'s handlers.
 ///
 /// [`Dispatcher`]: crate::dispatching::Dispatcher
+#[deprecated(note = "Use dispatching2 instead")]
 pub type DispatcherHandlerRx<R, Upd> = UnboundedReceiver<UpdateWithCx<R, Upd>>;