+++ src/lib.rs
@@ -6,17 +6,17 @@
 //!
 //! ([Full](https://github.com/teloxide/teloxide/blob/master/examples/dices_bot/src/main.rs))
 //! ```no_run
-//! use teloxide::prelude::*;
+//! use teloxide::prelude2::*;
 //!
 //! # #[tokio::main]
-//! # async fn main_() {
+//! # async fn main() {
 //! teloxide::enable_logging!();
 //! log::info!("Starting dices_bot...");
 //!
 //! let bot = Bot::from_env().auto_send();
 //!
-//! teloxide::repl(bot, |message| async move {
-//!     message.answer_dice().await?;
+//! teloxide::repls2::repl(bot, |message: Message, bot: AutoSend<Bot>| async move {
+//!     bot.send_dice(message.chat.id).await?;
 //!     respond(())
 //! })
 //! .await;
@@ -61,17 +61,26 @@
 // https://github.com/rust-lang/rust-clippy/issues/7422
 #![allow(clippy::nonstandard_macro_braces)]
 
-#[cfg(feature = "ctrlc_handler")]
 pub use dispatching::repls::{
     commands_repl, commands_repl_with_listener, dialogues_repl, dialogues_repl_with_listener, repl,
     repl_with_listener,
 };
 
+#[cfg(feature = "dispatching2")]
+pub use dispatching2::repls as repls2;
+
 mod logging;
 
+// Things from this module is also used for the dispatching2 module.
 pub mod dispatching;
+#[cfg(feature = "dispatching2")]
+#[cfg_attr(all(docsrs, feature = "nightly"), doc(cfg(feature = "dispatching2")))]
+pub mod dispatching2;
 pub mod error_handlers;
 pub mod prelude;
+#[cfg(feature = "dispatching2")]
+#[cfg_attr(all(docsrs, feature = "nightly"), doc(cfg(feature = "dispatching2")))]
+pub mod prelude2;
 pub mod utils;
 
 #[doc(inline)]
@@ -81,6 +90,9 @@ pub use teloxide_core::*;
 #[cfg_attr(all(docsrs, feature = "nightly"), doc(cfg(feature = "macros")))]
 pub use teloxide_macros as macros;
 
+#[cfg(feature = "dispatching2")]
+#[cfg_attr(all(docsrs, feature = "nightly"), doc(cfg(feature = "dispatching2")))]
+pub use dptree;
 #[cfg_attr(all(docsrs, feature = "nightly"), doc(cfg(feature = "macros")))]
 #[cfg(feature = "macros")]
 pub use teloxide_macros::teloxide;